perf(davinci): publish v-if branch keys during lowering (#5953)
