+++ crates/vize_s1_to_s2/src/lower/cx.rs
@@ -53,6 +53,7 @@ pub(crate) struct Cx<'a> {
     pub scopes: SideTable<ScopeFacts>,
     pub texts: SideTable<super::text::TextParts>,
     pub for_facts: SideTable<super::forop::ForParts>,
+    pub if_facts: SideTable<super::if_keys::IfFacts>,
     pub wrappers: SideTable<super::structural::WrapperKeys>,
     pub for_wrappers: SideTable<super::structural::ForWrapper>,
     pub caps: super::caps::LegacyCaps,
@@ -87,6 +88,7 @@ impl<'a> Cx<'a> {
             scopes: SideTable::new(),
             texts: SideTable::new(),
             for_facts: SideTable::new(),
+            if_facts: SideTable::new(),
             wrappers: SideTable::new(),
             for_wrappers: SideTable::new(),
             caps,
