+++ crates/vize_s1_to_s2/src/pass/legacy/ids.rs
@@ -62,6 +62,7 @@ pub(super) fn rekey(lowered: &mut Lowered<'_>, sync_ids: &[NodeId]) {
     }
     rekey_table(&mut lowered.scopes, sync_ids);
     rekey_table(&mut lowered.texts, sync_ids);
+    rekey_table(&mut lowered.if_facts, sync_ids);
     rekey_table(&mut lowered.for_facts, sync_ids);
     rekey_table(&mut lowered.wrappers, sync_ids);
     rekey_table(&mut lowered.for_wrappers, sync_ids);
