+++ crates/vize_s1_to_s2/tests/emit_unsupported_direct.rs
@@ -206,6 +206,7 @@ fn lowered<'a>(a: &'a Allocator, root: Region<'a>) -> Lowered<'a> {
         provenance: StdVec::new(),
         scopes: SideTable::new(),
         texts: SideTable::new(),
+        if_facts: SideTable::new(),
         for_facts: SideTable::new(),
         wrappers: SideTable::new(),
         for_wrappers: SideTable::new(),
