+++ davinci-road/plan/storage-inventory.tsv
@@ -110,12 +110,13 @@ s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/directive.rs	1	6	0	0	0	0	0	0
 s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/element.rs	1	2	1	1	1	4	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/lower/expr.rs	0	0	1	4	0	0	0	0
 s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/forop.rs	1	3	1	7	1	2	0	0
+s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/if_keys.rs	1	2	1	8	1	3	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/lower/html.rs	0	0	1	2	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/lower/leaf.rs	0	0	1	6	1	5	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/lower/once_memo.rs	0	0	1	3	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/lower/show.rs	0	0	1	2	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/lower/slot.rs	0	0	1	4	1	4	0	0
-s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/structural.rs	1	12	1	4	1	14	0	0
+s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/structural.rs	1	14	1	4	1	14	0	0
 s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/structural/wrapper.rs	1	7	1	15	0	0	0	0
 s1_to_s2	lower	crates/vize_s1_to_s2/src/lower/sugar.rs	1	1	1	2	0	0	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/lower/table.rs	0	0	0	0	1	11	0	0
@@ -135,8 +136,6 @@ s1_to_s2	-	crates/vize_s1_to_s2/src/pass/legacy/slot.rs	0	0	0	0	1	2	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/pass/legacy/sync.rs	0	0	1	2	1	5	0	0
 s1_to_s2	-	crates/vize_s1_to_s2/src/pass/legacy/tree.rs	0	0	0	0	1	1	0	0
 s1_to_s2	pass	crates/vize_s1_to_s2/src/pass/text.rs	1	1	0	0	0	0	0	0
-s1_to_s2	pass	crates/vize_s1_to_s2/src/pass/vif.rs	1	5	1	6	0	0	0	0
-s1_to_s2	-	crates/vize_s1_to_s2/src/pass/vif/keys.rs	0	0	1	4	1	3	0	0
 s1_to_s2	pass	crates/vize_s1_to_s2/src/pass/vmodel.rs	1	10	1	13	0	0	0	0
 s1_to_s2	pass	crates/vize_s1_to_s2/src/pass/vmodel/check.rs	1	1	1	5	0	0	0	0
 s1_to_s2	pass	crates/vize_s1_to_s2/src/pass/vslot.rs	1	2	1	4	0	0	0	0
