+++ tests/tooling/davinci-storage-inventory.ts
@@ -44,7 +44,7 @@ export const categoryReasons: Record<VecCategory, string> = {
 export const expectedProductionAllocVec: StorageSummary = {
   files: 77,
   directPaths: 89,
-  boundUses: 305,
+  boundUses: 304,
 };
 
 function count(value: string, line: number): number {
